+++ jsume/app.py
@@ -2,6 +2,6 @@
 
 from flask import Flask
 
-app = Flask("json_resume_to_pdf")
+app = Flask("jsume")
 
 LOGGER = app.logger
