+++ jsume/wsgi.py
@@ -1,7 +1,7 @@
 """WSGI entry point for the application."""
 
-from json_resume_to_pdf.app import app, LOGGER
-from json_resume_to_pdf.endpoints import *
+from jsume.app import app, LOGGER
+from jsume.endpoints import *
 import logging
