+++ setup.cfg
@@ -1,11 +1,11 @@
 
 
 [metadata]
-name = json-resume-to-pdf
+name = jsume
 version = file: version.txt
 author = Marco Pistilli
 author_email = pistilli.mrc@gmail.com
-url = https://github.com/marcopist/json-resume-to-pdf
+url = https://github.com/marcopist/jsume
 description = "An API to turn your resume.json gist into a resume.pdf"
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -29,7 +29,7 @@ classifiers =
 zip_safe = false
 include_package_data = true
 python_requires = >= 3.7
-packages = json_resume_to_pdf
+packages = jsume
 test_suite = tests
 setup_requires =
     setuptools
