+++ tests/endpoints/test_username.py
@@ -1,13 +1,13 @@
 from unittest.mock import patch
 from pytest import mark
 
-from json_resume_to_pdf.endpoints.username import route_username
+from jsume.endpoints.username import route_username
 
 from tests.utils.code.safe_get_gist import safe_get_gist
 from tests.utils.code.compare_pdf import compare_pdf
 
 
-@patch("json_resume_to_pdf.github.get_gist", safe_get_gist)
+@patch("jsume.github.get_gist", safe_get_gist)
 def test_route_username_not_published() -> None:
     """Test for json_resume.app.endpoints.route_username
     Assumes that no resume was published"""
@@ -16,7 +16,7 @@ def test_route_username_not_published() -> None:
     assert actual == ""
 
 
-@patch("json_resume_to_pdf.github.get_gist", safe_get_gist)
+@patch("jsume.github.get_gist", safe_get_gist)
 def test_route_username_invalid_schema() -> None:
     """Test for json_resume.app.endpoints.route_username
     Assumes that no resume was published"""
@@ -26,7 +26,7 @@ def test_route_username_invalid_schema() -> None:
     assert len(actual) > 0
 
 
-@patch("json_resume_to_pdf.github.get_gist", safe_get_gist)
+@patch("jsume.github.get_gist", safe_get_gist)
 def test_route_username_valid() -> None:
     """Test for json_resume.app.endpoints.route_username
     Assumes that a resume was published"""
